+++ packages/unlock-js/CHANGELOG.md
@@ -4,6 +4,11 @@
 
 Allow passing creator in the createLock method and improve typing.
 
+# 0.31.6
+
+- add `addLockManager` in walletService
+- add `renounceLockManager` in walletService
+
 # 0.31.5
 
 - Added suppport for `runEstimate`